lab2/user-service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5